Moved action menu on dataset page to standalone file. Change position to left of breadcrumb.
